Use /tmp/dunst-status for notifications state
As suggested in https://github.com/dunst-project/dunst/issues/216#issuecomment-421033384.

/tmp should be cleared on boot. Note that notifications is still turned on when starting i3.
That way we override whatever the default behaviour of dunst is (it defaults to on) and we
force the state to appear in i3status by creating the state file.

This is fine while waiting for dunstctl (https://github.com/dunst-project/dunst/pull/651).