NixOS/pulseaudio: Don't (try to) set default source
This did not have any effect anyway (not sure why), but I found that
pulseaudio automatically switches default source to headphones mic
when headphones are connected. Which makes sense, since then I'm muting
headphones mic when it is connected, and laptop mic when it is not.

In ALSA, Capture follows the PA headphone mute state and laptop mic
mute state does not seem to be followed by anything. The mic mute LEDs
works as follows (tested with Mic Mute-LED Mode set to Follow Mute):

-When either Capture or Dmic0 10 mute state changes, set LED to that
state
-On boot, LED is only lit when both Capture and Dmic0 is muted

As a result, mic mute LED works when headphones are connected, does
not work when they are not, and is potentially out of sync on boot.