picom: Backend on xr_glx_hybrid
* No issues with dropdown menus
* No screen tearing AFAICT
* Compatible with xsecurelock

The issues with dropdown menu items not being painted unless hovering
over them happens on the glx backend and, contrary to what I thought
earlier, does not seem to be related to the paint-on-overlay option. The
reason the issue appeared to be solved when using that option, is that
picom actually erred repeatedly because it will not start when using
deprecated options. Therefore, I assume, things looked right since picom
was constantly restarting and repainting. Obviously that is not the way
to go, and the issue with glx and dropdown menus persists.